Extracted ModelStub into it's own file
